Build report-ready table with all known plant species occurences for a region of interest (roi).

The tabulAve function will create a reference table (Excel file) considering all plant species occurring in a Roi, joining all relevant information (taxonomic, conservation, occurrence status and legal framework).

tabulFlora(
  utm_ae = utm_ae,
  utm_q = utm_contig,
  fielddata = NULL,
  biblio = "all"
)

Arguments

utm_ae

character vector of utm grid cells touching roi.

utm_q

character vector of utm grid cells around roi

fielddata

a dataframe with data obtained for the roi. Tipically a table with species and utm 10x10 code reference.

biblio

character vector. Databases to get data from. One of c('all', 'floraon', 'pterid').

ae

any spatial (sp, sf, spatialdataframe) or coordinates vector indicating a region of interest.

Value

a table with a complete and pre-formatted table, for use in reports.

Details

ensure that field datasheet follow the standardized formatting (Flora_TabulEIA_Template.xlsx).

Author

Paulo E. Cardoso

Examples

if (FALSE) {
# read roi from shapefile
library(tidyverse)
ae <- sf::read_sf(here::here("sig", "ae_buffer250m.shp"),
  stringsAsFactors = F
) %>%
  st_set_crs(3763)
# Cast multipolygon geometry to single parts
aeu <- st_cast(ae, "POLYGON")
# Add an UID to each
aeu$id <- c(1:2)
# get UTM codes for your roi
utm_all <- utm_id(grid = utm10k, roi = aeu %>% filter(id == 2), buff = NULL, contiguity = "queen")
# tabulate species occurence and status in the area
tflora <- tabulEIA::tabulFlora(
  fielddata = NULL, utm_ae = utm_ae, utm_q = utm_contig,
  biblio = "all"
)
# export to csv
write_excel_csv(tflora, path = here::here("output", "tabela_avifauna_agolada.csv"))
}