Build report-ready table with all known mammal occurences for a region of interest.

The tabulMam function will create a reference table (Excel file) considering all mammal species occuring in a region, joining all relevant information (taxonomic, conservation, occurence status and legal framework). It will also (in the future) summarize region status (ratios for protected and threatened species).

tabulMam(
  utm_ae = utm_ae,
  utm_q = utm_contig,
  fielddata = NULL,
  marinhos = FALSE,
  atlas = "all",
  dhab = dh13_18,
  inat = NULL
)

Arguments

utm_ae
character. vector of utm grid cells touching roi.

utm_q
character. vector of utm grid cells around roi

fielddata
dataframe. Table with bid data obtained for the roi. Tipically a table with species utm reference.

marinhos
Logical. False (default) exclude marine mammals. To include change to TRUE.

atlas
character. Mammal and Bat Atlas data with references to the UTM 10x10 grid cell.

dhab
character. Use Habitat Direective data from 2013-2018 report published by ICNF.

inat
character. Use Inaturalist Dataset for 2018 and 2019 with thousands of entries available on iNaturalist database from 2018 and 2019.

ae
sf object. any spatial (sp, spatialdataframe) or coordinates vector indicating a region of interest.

Value

an Excel file with a complete and pre-formatted table, (almost) ready for use on EIA reports.

Details

ensure that field datasheet follow the standardized formatting.

Author

Paulo E. Cardoso

Examples

   # read roi from shapefile
   ae = sf::read_sf(here::here('sig', 'ae_buffer250m.shp'),
                    stringsAsFactors = F) %>%
    st_set_crs(3763)

#> Error in st_set_crs(., 3763): could not find function "st_set_crs"

   # Cast multipolygon geometry to single parts
   aeu = st_cast(ae, "POLYGON")

#> Error in st_cast(ae, "POLYGON"): could not find function "st_cast"

   # Add an UID to each
   aeu$id = c(1:2)

#> Error in aeu$id = c(1:2): object 'aeu' not found

   # get UTM codes for your roi
   utm_all = utm_id(grid = utm10k, roi = aeu %>% filter(id == 2), buff = NULL, contiguity = 'queen')

#> Error in filter(., id == 2): object 'aeu' not found

   # tabulate species occurence and status in the area
   tmam = tabulEIA::tabulMam(utm_ae = utm_ae, utm_q = utm_contig, fielddata = NULL, atlas = 'all', inat = FALSE)

#> Error: Problem with `filter()` input `..1`.
#> x object 'utm_ae' not found
#> i Input `..1` is `utm %in% c(utm_ae, utm_q)`.

   # export to csv